피드백 반영 : domain -> view 의존 끊기
RaceLog 클래스 신규 작성
경기 결과를 raceLog에 담아 전달하면 view에서 이를 읽어 출력
