+++ src/main/kotlin/racingCar/domain/RaceLog.kt
@@ -1,0 +1,10 @@
+package racingCar.domain
+
+class RaceLog {
+    var carsLog: MutableList<List<Car>> = mutableListOf()
+    lateinit var winners: List<Car>
+
+    fun addLog(carList: List<Car>) {
+        carsLog += carList
+    }
+}
