+++ src/main/kotlin/racingCar/view/DisplayView.kt
@@ -1,6 +1,7 @@
 package racingCar.view
 
 import racingCar.domain.Car
+import racingCar.domain.RaceLog
 
 
 object DisplayView {
@@ -16,9 +17,19 @@ object DisplayView {
     }
 
     /**
-     * 한 번 진행이 될 때마다 자동차의 상태를 화면에 출력한다.
+     * 경기 결과를 화면에 출력한다.
      */
-    fun displayRaceProgress(carList: List<Car>) {
+    fun displayRaceResult(raceResult: RaceLog) {
+        raceResult.carsLog.forEach{ displayRaceProgress(it) }
+
+        val winners = raceResult.winners.joinToString(WINNER_DELIMITER) { it.name }
+        println("${winners}가 최종 우승했습니다.")
+    }
+
+    /**
+     * 경기 진행에 따른 자동차의 상태를 화면에 출력한다.
+     */
+    private fun displayRaceProgress(carList: List<Car>) {
         carList.forEach { displayCar(it) }
         println()
     }
@@ -26,11 +37,6 @@ object DisplayView {
     /**
      * 우승자를 화면에 출력한다.
      */
-    fun displayRaceResult(carList: List<Car>) {
-        val winners = carList.joinToString(WINNER_DELIMITER) { it.name }
-        println("${winners}가 최종 우승했습니다.")
-    }
-
     private fun displayCar(car: Car) {
         println(car.name + NAME_DELIMITER + CAR_POSITION_MARK.repeat(car.position))
     }
